feat: :sparkles: change sqlite to postgres database and some features

+++ src/entity/Geolocation.ts
@@ -5,9 +5,9 @@ export class Geolocation {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column('double')
+  @Column('float')
   latitute: number;
 
-  @Column('double')
+  @Column('float')
   longitude: number;
 }
